Clean up code in Spark, and replace current hlsl to Blur hlsl

+++ Framework/AI/Src/FleeBehavior.cpp
@@ -7,7 +7,7 @@ JimmyGod::Math::Vector2 FleeBehavior::Calculate(Agent & agent)
 {
 	if (IsActive())
 	{
-		const double PanicDistanceSq = panicDistance * panicDistance;
+		const float PanicDistanceSq = panicDistance * panicDistance;
 		if (JimmyGod::Math::DistanceSqr(agent.Position, agent.Destination) > PanicDistanceSq)
 			return JimmyGod::Math::Vector2{ 0,0 };
 		if (IsDebugUIActive())
